Refactor use of FishSwimming to its own component
